move pool candidate fragment to own file
